capstone-projects/option_a_currency_exchange/starter_code/requirements.txt
# Core dependencies
google-adk>=1.0.0
google-generativeai>=0.8.0
google-cloud-aiplatform>=1.70.0

# MCP support
mcp>=1.0.0
fastmcp>=0.4.0

# Web server for A2A
uvicorn>=0.30.0
fastapi>=0.111.0

# HTTP client
httpx>=0.27.0
aiohttp>=3.9.0

# Environment management
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Development
black>=24.0.0
isort>=5.13.0
mypy>=1.8.0
